Add a Crash service that detects whether a user is in a car crash using the accelerometer sensor and the ActivityTransitionAPI
The values from the accelerometer sensor, combined with the ActivityTransitionAPI will be used to detect whether a user is in a car crash or not. If the current activity detected by the API is IN_VEHICLE, and if the force of shaking given to device is above a certain threshold, then it is most likely that the user encountered a crash.